Delete Project Logic moved
