+++ packages/frontend-2/components/server-management/DeleteProjectDialog.vue
@@ -1,5 +1,10 @@
 <template>
-  <LayoutDialog v-model:open="isOpen" max-width="sm" :title="title" :buttons="buttons">
+  <LayoutDialog
+    v-model:open="isOpen"
+    max-width="sm"
+    :title="title"
+    :buttons="dialogButtons"
+  >
     <div class="flex flex-col gap-6">
       <p>
         Are you sure you want to
@@ -26,20 +31,77 @@
 import { computed } from 'vue'
 import { LayoutDialog } from '@speckle/ui-components'
 import { ProjectItem } from '~~/lib/server-management/helpers/types'
+import { useMutation } from '@vue/apollo-composable'
+import {
+  convertThrowIntoFetchResult,
+  getFirstErrorMessage
+} from '~~/lib/common/helpers/graphql'
+import { useGlobalToast, ToastNotificationType } from '~~/lib/common/composables/toast'
+import { graphql } from '~~/lib/common/generated/gql'
+
+const adminDeleteProject = graphql(`
+  mutation AdminPanelDeleteProject($ids: [String!]) {
+    streamsDelete(ids: $ids)
+  }
+`)
+
+const { triggerNotification } = useGlobalToast()
+const { mutate: adminDeleteMutation } = useMutation(adminDeleteProject)
 
 const emit = defineEmits<{
   (e: 'update:open', val: boolean): void
+  (e: 'project-deleted', val: string): void
 }>()
 
 const props = defineProps<{
   title: string
   open: boolean
   project: ProjectItem | null
-  buttons?: Array<{ text: string; props: Record<string, unknown>; onClick: () => void }>
 }>()
 
 const isOpen = computed({
   get: () => props.open,
   set: (newVal) => emit('update:open', newVal)
 })
+
+const deleteConfirmed = async () => {
+  const projectId = props.project?.id
+  if (!projectId) {
+    return
+  }
+
+  const result = await adminDeleteMutation({
+    ids: [projectId]
+  }).catch(convertThrowIntoFetchResult)
+
+  if (result?.data?.streamsDelete) {
+    triggerNotification({
+      type: ToastNotificationType.Success,
+      title: 'Project deleted',
+      description: 'The project has been successfully deleted'
+    })
+    emit('project-deleted', projectId)
+    emit('update:open', false)
+  } else {
+    const errorMessage = getFirstErrorMessage(result?.errors)
+    triggerNotification({
+      type: ToastNotificationType.Danger,
+      title: 'Failed to delete project',
+      description: errorMessage
+    })
+  }
+}
+
+const dialogButtons = [
+  {
+    text: 'Delete',
+    props: { color: 'danger', fullWidth: true },
+    onClick: deleteConfirmed
+  },
+  {
+    text: 'Cancel',
+    props: { color: 'secondary', fullWidth: true, outline: true },
+    onClick: () => emit('update:open', false)
+  }
+]
 </script>
